backend/requirements.txt
torch==2.1.0
torchvision==0.16.0
diffusers==0.24.0
transformers==4.35.0
Pillow==10.1.0
numpy==1.24.3
tqdm==4.66.1
opencv-python==4.8.1.78
requests==2.31.0
python-dotenv==1.0.0
dashscope==1.13.6
oss2==2.18.3
python-multipart==0.0.6
aiohttp==3.9.1 